backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.23
motor>=3.3
pydantic>=2.6
python-dotenv>=1.0
orjson>=3.9
dnspython>=2.6
passlib[bcrypt]>=1.7
python-jose[cryptography]>=3.3
pytest>=8.0
